Update README.md
Neural networks are powerful tools in machine learning given the high precision rates obtained in high dimensional problems. In this session we cover the basics of neural networks. This session includes a brief description of what are networks, and what are the types of networks commonly used in machine learning (convolutional, recurrent, LSTM, among others). In the practical example, we construct a neural network for a regression problem from scratch. For this, we construct the traditional cross-entropy loss function, we present the main types of activation functions and the optimization routines that are more frequent in this type of problems